pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "drf-spectacular>=0.29.0",
    "ruff>=0.15.9",
]
prod = [
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
]

[tool.ruff]
line-length = 88  # black과 호환
target-version = "py312"  # Python 3.12 가정 (Django 4.2 지원)
exclude = ["migrations"]  # Django 마이그레이션 파일 제외

[tool.ruff.lint]
select = [
    "E",  # pycodestyle 오류
    "F",  # pyflakes
    "I",  # isort (임포트 정렬)
    "DJ", # Django-specific 규칙
    "UP", # pyupgrade (최신 Python 문법 적용)
    "B",  # bugbear (잠재적 버그 검사)
    "Q",  # flake8-quotes (따옴표 스타일 강제)
]
ignore = ["E203", "F403", "F405"]  # black과 호환성 위해
fixable = ["ALL"]  # 모든 수정 가능한 문제 자동 수정

[tool.ruff.format]
quote-style = "double"  # black과 일치
indent-style = "space"
skip-magic-trailing-comma = false
